# internal/adapters/sqlite/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
)

type Absence struct {
	ID       int64
	Date     string
	Type     string
	Fraction float64
	Note     sql.NullString
}

type Company struct {
	ID        int64
	Name      string
	CreatedAt int64
}

type Config struct {
	Key   string
	Value string
}

type Entry struct {
	ID      int64
	Kind    string
	StartTs int64
	EndTs   sql.NullInt64
	Note    sql.NullString
	TaskID  sql.NullInt64
}

type EntryProject struct {
	EntryID   int64
	ProjectID int64
}

type Project struct {
	ID             int64
	Name           string
	Archived       int64
	CreatedAt      int64
	Color          sql.NullString
	Note           sql.NullString
	CompanyID      sql.NullInt64
	JiraProjectKey sql.NullString
}

type Task struct {
	ID        int64
	ProjectID int64
	JiraKey   sql.NullString
	Title     string
	Archived  int64
	CreatedAt int64
}